StickyComments: Add simple first implementation.
When a comment thread loads, the plugin will look for a distinguished comment.
If one is found (regardless of depth), it is pulled out and inserted at the
top of the page.  An element is left in its original place to mark where it
came from, and its children are all collapsed.